requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pandas==2.2.3
plotly==5.18.0
psycopg2-binary>=2.9.9
